WIP: Catch fatal Rust errors in Java
This works for all but open/closeBackend. This is because we currently return
raw values rather than protobuf byte streams.

This is WIP as the Rust could be implemented better - possibly a macro

A function was attempted and should be tried again with a single lambda
Multiple parameters made execution significantly slower.

But, learning how to do macros can wait until the library is released.